Search auth-source for api key before authn (#110)
* search auth-sources for api-key before authn

This commit closes https://linear.app/mercury/issue/84

**Why:**
For users who have saved their api-key in an auth-source, this commit
prevents them from seeing the prompt to authenticate manually or
automatically, which can be annoying and takes time.

**How:**
implement codeium-get-saved-api-key using hardcoded values for :host and
for :user. I'm not sure if it makes sense to allow the user to pass in
arguments instead of hard-coding values here, but personally I think
this is ok since I have seen other packages do the same. As an example,
see
https://github.com/rksm/org-ai/blob/812b59f88851536ed3ded199fb55d303c16c7c12/org-ai-openai.el#L190-L191

* Add instructions for using auth-sources to authn

https://linear.app/mercury/issue/84